refactor(launcher): tidy up SimpleLauncher app entry point [ARTP-976]
